refactor(oca-ast): add recursive_attributes submodule
